+++ theano/tensor/nnet/blocksparse.py
